Move `ImageCommentsViewController` helper methods to separate file
